Updating the SERVICES module with more services.
Updated from the bottom up all widgets with all corresponding units. Updated new "services" components to make the unit conversion on the fly as soon as the user switches units. The advantage of the method use to update the services is that now any widget that allows conversion will be able to be re-used anywhere in STA.

git-svn-id: file:///home/patrick/Documents/STA_svn/trunk@769 0f720082-229d-4b06-8c75-e5250cb1739c